api/requirements.txt
flask==3.0.0
google-cloud-vision==3.7.2
werkzeug==3.0.1
duckduckgo-search==5.3.0
google-generativeai==0.3.2
